Transaction

+++ src/main/java/pruebaTecnica/bookStore/dto/response/transaction/TransactionResponse.java
@@ -17,7 +17,7 @@ public class TransactionResponse {
 
     public List<String> sales;
 
-    public int totalRevenue;
+    public double totalRevenue;
 
     public List<String> customers;
 }
